notification_service/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via pytest
backports-zoneinfo==0.2.1
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   pip-compile-multi
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
    #   drf-yasg
coverage[toml]==6.3.2
    # via pytest-cov
django==4.0.3
    # via
    #   -r requirements.in
    #   django-annoying
    #   django-cors-headers
    #   django-js-asset
    #   django-ranged-response
    #   django-simple-captcha
    #   django-taggit
    #   djangorestframework
    #   drf-yasg
django-annoying==0.10.6
    # via -r requirements.in
django-ckeditor==6.3.1
    # via -r requirements.in
django-cors-headers==3.11.0
    # via -r requirements.in
django-js-asset==2.0.0
    # via
    #   -r requirements.in
    #   django-ckeditor
django-ranged-response==0.2.0
    # via
    #   -r requirements.in
    #   django-simple-captcha
django-simple-captcha==0.5.17
    # via -r requirements.in
django-taggit==2.1.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.1
    # via
    #   -r requirements.in
    #   coreschema
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via
    #   drf-yasg
    #   pytest
pep517==0.12.0
    # via pip-tools
phonenumberslite==8.12.46
    # via -r requirements.in
pillow==9.1.0
    # via django-simple-captcha
pip-compile-multi==2.4.5
    # via -r requirements.in
pip-tools==6.6.0
    # via pip-compile-multi
pluggy==1.0.0
    # via pytest
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
requests==2.27.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via
    #   django-annoying
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   coverage
    #   pep517
    #   pytest
toposort==1.7
    # via pip-compile-multi
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
